fix: moved NotificationMarkdown to molecules
