+++ src/components/molecules/NotificationMarkdown/index.ts
@@ -1,0 +1,1 @@
+export {default} from './NotificationMarkdown';
